Move agent regions code to own submodule
